% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{aggEmbryos}
\alias{aggEmbryos}
\title{aggEmbryos merge a list of CD-like dataFrame into one dataFrame, with options to keep embryo identifier}
\usage{
aggEmbryos(datList, labelEmbryo = TRUE)
}
\arguments{
\item{datList}{list of CD-like dataFrames}

\item{labelEmbryo}{Boolean, whether to add a column to identify the embryo each cell come from}
}
\value{
a dataframe combining all CD-like dataframes in the first dimension
}
\description{
aggEmbryos merge a list of CD-like dataFrame into one dataFrame, with options to keep embryo identifier
}
