% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{blotAvgFun}
\alias{blotAvgFun}
\title{blotAvgFun}
\usage{
blotAvgFun(datList, cutoff = 0.5)
}
\arguments{
\item{datList}{a list of embryo cell tracking tables}

\item{cutoff}{the minimum percentile of blot values to consider (account for the underestimations during cell division)}
}
\description{
get the average blot value for each cell, can choose a cutoff to select for top blot values only
}
