% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{blotRange}
\alias{blotRange}
\title{blotRange return the range (min, max) of blot values in a given CD-like dataframe}
\usage{
blotRange(datList)
}
\arguments{
\item{datList}{a list of embryo cell tracking tables}
}
\value{
a named list with a value 'max' and a value 'min'
}
\description{
blotRange return the range (min, max) of blot values in a given CD-like dataframe
}
