% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{CD_In}
\alias{CD_In}
\title{CD_In:}
\usage{
CD_In(
  directory,
  prefix = "",
  suffix = ".csv",
  info.prefix = "",
  info.suffix = NULL,
  time.prefix = "",
  time.suffix = NULL
)
}
\arguments{
\item{directory}{directory the directory (folder) directly embryo cell tracking files}

\item{prefix}{prefix to identify embryo cell tracking files}

\item{suffix}{suffix (including file format) to identify embryo cell tracking files}

\item{info.prefix}{prefix to identify embryo info files}

\item{info.suffix}{suffix (including file format) to identify embryo info files}

\item{time.prefix}{prefix to identify embryo imaging time record files}

\item{time.suffix}{suffix (including file format) to identify embryo imaging time record files}
}
\value{
a large list made of:
CD: a list of Dataframes created by reading each embryo cell tracking file
info: a list of Dataframes each recoding informations of corresponding embryos
}
\description{
Function to batch import embryo data
}
