% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{CD_Save}
\alias{CD_Save}
\title{CD_Save}
\usage{
CD_Save(CDDataL, saveDir)
}
\arguments{
\item{CDDataL}{the list of CD Data generated by function \code{CD_In}}

\item{saveDir}{directory to output the tables and info file}
}
\description{
save cell tracking files along with theembryo informations
}
