% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plots.R
\name{CD_tree_plot}
\alias{CD_tree_plot}
\title{CD_tree_plot}
\usage{
CD_tree_plot(
  CD,
  root = "P0",
  min_gain = -1000,
  max_gain = FALSE,
  transform = "identity",
  start_time = FALSE,
  end_time = FALSE,
  plotExpression = TRUE,
  lowerBound = 0.4,
  upperBound = 0.97,
  tip_lab = TRUE,
  node_lab = FALSE,
  exp_col = "blot",
  cell_col = "cell",
  time_col = "time",
  extra_branch_l = 15,
  exp_legend = "Expression"
)
}
\arguments{
\item{CD}{dataframe for the CD table}

\item{root}{root cell of the tree}

\item{min_gain}{for color scale, the value corresponding to "minimum" color}

\item{max_gain}{for color scale, the value corresponding to "maximum" color}

\item{transform}{color scale transformation to apply}

\item{start_time}{only consider data before \code{start_time}}

\item{end_time}{only consider data before \code{end_time}}

\item{plotExpression}{whether to color each branch by the average expression of the cell}

\item{lowerBound}{lower bound of each cell's expression to count into average expression value}

\item{upperBound}{upper bound of each cell's expression to count into average expression value}

\item{tip_lab}{Boolean, whether to label the tree "leaves" with cell names}

\item{node_lab}{Boolean, whether branch poins with cell names}

\item{exp_col}{name of the column that correspond to expression attributes}

\item{cell_col}{name of the column that correspond to cell names attributes}

\item{time_col}{name of the column that correspond to cell time attributes}

\item{extra_branch_l}{length of dummy branches to connect trees to root P0 or P if such cell were not tracked}

\item{exp_legend}{name of the color scale}
}
\value{
ggplot figure of the tree plot
}
\description{
tree plots for a list of embryo cell tracking data
}
