% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{dataCorrection}
\alias{dataCorrection}
\title{dataCorrection}
\usage{
dataCorrection(
  CDDataL,
  cell = NULL,
  lineage = "P0",
  alignAt = NaN,
  model,
  exc_zMin,
  exc_zMax,
  zMax
)
}
\arguments{
\item{CDDataL}{the list of CD Data generated by function \code{CD_In}}

\item{cell}{only correct some specific cells}

\item{lineage}{only correct some specific lineages}

\item{alignAt}{the cell used to find common time point to align}

\item{model}{the model generated by \code{depthCorrectionParm}}

\item{exc_zMin}{excitation at z=1}

\item{exc_zMax}{excitation at maximum z}

\item{zMax}{maximum z}
}
\value{
corrected \code{CDDataL}
}
\description{
apply the data correction parameters generated by \code{depthCorrectionParm} to do correction
}
