% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{depthCorrectionParm}
\alias{depthCorrectionParm}
\title{depthCorrectionParm:}
\usage{
depthCorrectionParm(
  CDDataL,
  cell = NULL,
  lineage = NULL,
  alignCell,
  startT = 0,
  endT = Inf,
  exc_zMin,
  exc_zMax,
  zMax
)
}
\arguments{
\item{CDDataL}{the list of CD Data generated by function \code{CD_In}}

\item{startT}{time after alignment (as frame number), only data after this time point will be considered in fitting model}

\item{endT}{time after alignment (as frame number), only data before this time point will be considered}

\item{exc_zMin}{excitation at z=1}

\item{exc_zMax}{excitation at maximum z}

\item{zMax}{maximum z}

\item{cells}{array of cell names to be considered for depth correction}

\item{lineages}{array of lineages mother cell names to be considered for depth correction}

\item{alignAt}{the cell name by which all embryo time will be aligned at its end, and its end time will become 0 after alignment}
}
\value{
the regression model from fitting
}
\description{
get the depth (\code{z}) correction parameters based on the sample data with an exponential model: \cr
\code{blot = blotReal * (exp(a*z) * b)} (1) for \code{a} and \code{b} are factors to be determined\cr
\code{blot/blotReal = exp(a*z) * b} (2) \cr
\code{ln(blot/blotReal) = a*z + b} (3) \cr
Assuming the real blot is the same across cells at the same timepoint (after aligning development stages), we
will make the mean of the same cell for each timepoint the \code{blotReal}
}
