% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEmb.R
\name{drawEmbLine}
\alias{drawEmbLine}
\title{drawEmbLine}
\usage{
drawEmbLine(
  embryoCD,
  time,
  lineages = NULL,
  title = NULL,
  colors = NULL,
  alphas = NULL,
  outline = "#444",
  xSize = 1,
  ySize = 1,
  zSize = 1,
  aligned = F,
  alpha_other = 0.2,
  cellSize = 7.5,
  center = list(x = 0, y = 0, z = 0),
  viewPoint = list(x = 0, y = 0, z = 1.8)
)
}
\arguments{
\item{embryoCD}{the embryo dataframe with cell, time column, mandatory}

\item{time}{numeric, mandatory, which time to plot the embryo, automatically choose the closest time point if no exact match}

\item{lineages}{a list of lineages to highlight (plot with the given color and opacity in "colors" and "alphas" parameter)}

\item{title}{figure title text}

\item{colors}{a list of colors to plot each lineage given in "lineages"}

\item{alphas}{a list of alpha i.e. opacity for each lineage given in "lineages"}

\item{outline}{an R color value to specify the outline color of nucleus
\itemize{
\item default as gray outline
\item input \code{NULL} to have no pixels
\item outline have a fixed width by pixel, if the outline covers the whole area of nucleus marker, try increase \code{cellSize} or turn off outline
}}

\item{xSize}{size of each x unit in embryoCD dataframe, default 1}

\item{ySize}{size of each y unit in embryoCD dataframe, default 1}

\item{zSize}{size of each z unit in embryoCD dataframe, default 1}

\item{alpha_other}{opacity for cells not specified}

\item{cellSize}{single numeric, size of the data points}

\item{center}{A named list with elements:
\describe{
\item{x}{Numeric. X coordinate.}
\item{y}{Numeric. Y coordinate.}
\item{z}{Numeric. Z coordinate.}
}
which point in the 3d space does the camera aim at, default list(x=0,y=0,z=0)}

\item{viewPoint}{A named list with elements:
\describe{
\item{x}{Numeric. X coordinate.}
\item{y}{Numeric. Y coordinate.}
\item{z}{Numeric. Z coordinate.}
}
where to put the imaginative camera initially, default list(x=0,y=0,z=1.8)}
}
\value{
a list made of: the plotly figure, a dataframe of the highlighted cells, and a dataframe of cells not highlighted
}
\description{
draw the nucleus positions of the embryo in 3D scatter plots, colored by lineage
}
