% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{Emb_drop}
\alias{Emb_drop}
\title{Emb_drop}
\usage{
Emb_drop(CDDataL, dropEmb = NaN)
}
\arguments{
\item{CDDataL}{the list generated by \code{CD_In}}

\item{dropEmb}{names of embryos todrop}
}
\value{
dropped \code{CDDataL}
}
\description{
Remove an embryo data from the list generated by \code{CD_In}
}
