% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{findCell}
\alias{findCell}
\title{findCell}
\usage{
findCell(CDDataL, cellName)
}
\arguments{
\item{CDDataL}{list returned by \code{CD_In}}

\item{cellName}{name of the cell}
}
\value{
a list of dataframes recording all instances of the cell';s occurrence
}
\description{
return the occurrence of a cell name from the list generated by \code{CD_In}
}
