% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{getCells}
\alias{getCells}
\title{getCells}
\usage{
getCells(
  CDDataL,
  includeCD = NaN,
  excludeCD = NULL,
  cells = NULL,
  lineages = NULL,
  times = "ALL",
  dataReturn = TRUE
)
}
\arguments{
\item{CDDataL}{the list of CD Data generated by function \code{CD_In}}

\item{includeCD}{the embryos/CD file name to be extracted}

\item{excludeCD}{the embryos/CD file name to be avoid (will override \code{include})}

\item{cells}{(list of) exact cell names to extract}

\item{lineages}{(list of) mother cell name of lineage to extract}

\item{dataReturn}{whether this function return the data (TRUE) or index (FALSE) of selected cells}
}
\value{
a list of dataframes with the cell data, indexed by CD file embryo names
}
\description{
extract the data of given cells from each CD file
}
