% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{grepCells}
\alias{grepCells}
\title{grepCells}
\usage{
grepCells(
  CDData,
  cells = NULL,
  lineages = NULL,
  times = "ALL",
  timesDiffThreshold = 1.5,
  dataReturn = TRUE
)
}
\arguments{
\item{CDData}{a CD-like dataFrame}

\item{cells}{cells to retrieve}

\item{lineages}{mother cell of lineages to retrieve (inclusive to mother)}

\item{times}{scope of time to select from}

\item{timesDiffThreshold}{define how much difference between the input \code{time} and retrieved \code{time} is allowed, if there are no exact \code{time} match, the function will look for closest \code{time} value}
}
\value{
if \code{dataReturn==TRUE}: dataFrame of selected cells, if \code{dataReturn==FALSE}: return the row index in the original \code{CDData}
}
\description{
extract the data of given cells from the given CD table
}
