% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{lineageNames}
\alias{lineageNames}
\title{lineageNames}
\usage{
lineageNames(lineage)
}
\arguments{
\item{lineage}{a string or list of lineage name / mother cell name}
}
\value{
a list of regular expressions for cell
}
\description{
returns all possible starts of daughter lineages (including the mother cell itself) under a given lineage/mother cell
}
\examples{
lineageNames("EMS")
}
