% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{lineageRE}
\alias{lineageRE}
\title{lineageRE returns all regular expressions needed to grep cells under a given lineage/mother cell}
\usage{
lineageRE(lineage)
}
\arguments{
\item{lineage}{a string or list of lineage name / mother cell name, x as single-letter wildcard}
}
\value{
a list of regular expressions for cell
}
\description{
lineageRE returns all regular expressions needed to grep cells under a given lineage/mother cell
}
\examples{
lineageRE(c("EMS", "C"))
}
