% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{OrthgonalProj}
\alias{OrthgonalProj}
\title{Gram-Schmidt Process gives the orthogonal component to AP axis}
\usage{
OrthgonalProj(v, u)
}
\description{
Gram-Schmidt Process gives the orthogonal component to AP axis
}
