% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plots.R
\name{plot_trees}
\alias{plot_trees}
\title{plot_trees}
\usage{
plot_trees(
  inDir,
  root = "P0",
  plotExpression = T,
  lowerBound = 0.4,
  upperBound = 0.97,
  min_gain = 100,
  max_gain = 10000,
  transform = "identity",
  start_time = FALSE,
  end_time = FALSE,
  tip_lab = TRUE,
  node_lab = FALSE,
  outDir = NaN,
  exp_col = "blot",
  cell_col = "cell",
  time_col = "time",
  width = 7,
  height = 7,
  dpi = 300
)
}
\arguments{
\item{inDir}{directory of embryo cell tracking files to plot into trees}

\item{root}{root cell of the tree}

\item{plotExpression}{whether to color each branch by the average expression of the cell}

\item{lowerBound}{lower bound of each cell's expression to count into average expression value}

\item{upperBound}{upper bound of each cell's expression to count into average expression value}

\item{min_gain}{for color scale, the value corresponding to "minimum" color}

\item{max_gain}{for color scale, the value corresponding to "maximum" color}

\item{transform}{color scale transformation to apply}

\item{start_time}{only consider data before \code{start_time}}

\item{end_time}{only consider data before \code{end_time}}

\item{tip_lab}{Boolean, whether to label the tree "leaves" with cell names}

\item{node_lab}{Boolean, whether branch poins with cell names}

\item{outDir}{output directory}

\item{exp_col}{name of the column that correspond to expression attributes}

\item{cell_col}{name of the column that correspond to cell names attributes}

\item{time_col}{name of the column that correspond to cell time attributes}

\item{width}{in inch, figure width}

\item{height}{in inch, figure height}

\item{dpi}{dpi of output figure}
}
\value{
\code{NA}
}
\description{
tree plots for a list of embryo cell tracking data
generate one file for each CD file in the given directory, with consistent parameters
}
\details{
This function uses \link[ggtree:ggtree]{ggtree::ggtree} (Bioconductor) to plot trees.
#' The \code{ggtree} package must be installed separately:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{if (!requireNamespace("BiocManager", quietly = TRUE))
    install.packages("BiocManager")
BiocManager::install("ggtree")
}\if{html}{\out{</div>}}
}
