% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{readEmbryoTable}
\alias{readEmbryoTable}
\title{LineageProcessing
This file contains several utility functions that helps processing data in CD-like dataFrames
readEmbryoTable}
\usage{
readEmbryoTable(file, time.file = NULL)
}
\arguments{
\item{file}{file directory of the embryo cell tracking table}

\item{time.file}{embryo imaging time record files}
}
\value{
a list of: \verb{[[1]]} the embryo cell tracking table \verb{[[2]]} Boolean, if a time file is loaded
}
\description{
load a single embryo cell tracking file into data.framne
}
