% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEmb.R
\name{saveEmbImg}
\alias{saveEmbImg}
\title{printEmbImg Generates a .png image from drawEmbExp or drawEmbLine output figure
Require Reticulate that connects to a python distribution with "kaleido" and "plotly" installed
recommend setting up a python venv for this and specify the python kernel by:
reticulate::use_virtualenv("the/directory/holding/python/executable")}
\usage{
saveEmbImg(
  fig,
  output_file = "embryo_DV_view.png",
  center = list(0, 0, 0),
  viewPoint = list(x = 0, y = 0, z = 1.8),
  up = list(x = 0, y = 1, z = 0),
  width = 600,
  height = 450,
  keepLabels = FALSE,
  zoom = 1.3,
  pad_frac = 0.07,
  projection = c("perspective", "orthographic")
)
}
\arguments{
\item{fig}{plotly figure}

\item{output_file}{output file name}

\item{center}{coordinate the camera face}

\item{viewPoint}{from where does the camera look at the embryo}

\item{up}{which direction (as unit vector) the upper side of the camer points to}

\item{width}{in pixels, width of output image}

\item{height}{in pixels, height of output image}

\item{keepLabels}{Boolean, to keep the color bar and legends or not}

\item{zoom}{zoom factor for rendering}

\item{pad_frac}{how much to pad the axis ranges to avoid edge clash}

\item{projection}{either "orthographic" or "perspective"}
}
\description{
printEmbImg Generates a .png image from drawEmbExp or drawEmbLine output figure
Require Reticulate that connects to a python distribution with "kaleido" and "plotly" installed
recommend setting up a python venv for this and specify the python kernel by:
reticulate::use_virtualenv("the/directory/holding/python/executable")
}
