% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{timeAvg}
\alias{timeAvg}
\title{timeAvg}
\usage{
timeAvg(datList, attribute)
}
\arguments{
\item{datList}{list of CD-like dataFrames}

\item{attribute}{the attribute to be averaged}
}
\value{
dataFrame,
}
\description{
averages the selected attribute for all cells of the same time, can add corresponding standard deviation attribute
}
