% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{totalRePosition}
\alias{totalRePosition}
\title{totalRePosition}
\usage{
totalRePosition(
  CDFrame,
  time,
  indicatorP = "C",
  indicatorD = "Cxa",
  indicatorV = "MSxxp",
  indicatorL = NULL,
  indicatorR = NULL,
  xSize = 0.08651,
  ySize = 0.08651,
  zSize = 0.5
)
}
\arguments{
\item{CDFrame}{a CD-like dataFrame}

\item{time}{the time to generate rotation parameters, nucleus across all time points will be rotated with the same parameter}

\item{indicatorP}{lineage on the relative posterior of embryo center}

\item{indicatorD}{lineage on the relative dorsal of embryo center, use NULL not NaN if not considering}

\item{indicatorV}{lineage on the relative ventral of embryo center, use NULL not NaN if not considering}

\item{indicatorL}{lineage on the relative left of embryo center, use NULL not NaN if not considering}

\item{indicatorR}{lineage on the relative right of embryo center, use NULL not NaN if not considering}

\item{xSize}{size of each unit on x-axis in original CDFrame}

\item{ySize}{size of each unit on y-axis in original CDFrame}

\item{zSize}{size of each unit on z-axis in original CDFrame}
}
\value{
the provided CDFrame rotated
}
\description{
center the embryo at 0,0,0 coordinate and rotate the embryo across all time points
this uses the same strategy as \code{RePosition} and defines the orientation based on the specified
time point, then applies the same rotation to all time points
}
