% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{trimCD}
\alias{trimCD}
\title{trimCD}
\usage{
trimCD(CDDataL, drops)
}
\arguments{
\item{CDDataL}{the list of CD Data generated by function \code{CD_In}}

\item{drops}{names of columns to drop}
}
\value{
the CDDataL
}
\description{
drop the provided columns from all imported CD data
}
