% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEmb.R
\name{AddGroupExp}
\alias{AddGroupExp}
\title{AddGroupExp}
\usage{
AddGroupExp(
  plotlyFig,
  groupName,
  data,
  selectedCells,
  symbol,
  colorScheme,
  colorMin,
  colorMax,
  cellSize,
  opacity = 1
)
}
\value{
modified plotly figure
}
\description{
auxiliary function for \code{drawEmbVal}
}
