% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEmb.R
\name{AddGroupLine}
\alias{AddGroupLine}
\title{AddGroupLine}
\usage{
AddGroupLine(
  plotlyFig,
  groupName,
  data,
  selectedCells,
  color,
  opacity,
  cellSize
)
}
\value{
modified plotly figure
}
\description{
auxiliary function for \code{drawEmbLine}
}
