% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{alignTime}
\alias{alignTime}
\title{alignTime}
\usage{
alignTime(datList, alignCell, alignPoint = 1, align_t = 0, alignBlot = F)
}
\arguments{
\item{datList}{list of CD-like dataFrames}

\item{alignCell}{the cell used to find common time point to align}

\item{alignPoint}{a 0 to 1 ratio number showing what ratio of the alignCell life to align together (1 means the end, 0 means the start)}

\item{align_t}{the time of the \code{alignCell} at \code{alignPoint} after alignment, if the input is \code{"mean"} the align time will be the mean value of \code{alignPoint} across all embryos}

\item{alignBlot}{whether and method to align the blot values together default \code{FALSE} for not align, "mean" for aligning the mean values together (by addition/subtraction)}
}
\value{
modified list of CD-like dataFrames
}
\description{
'align' the cell time of a list of CD-like dataFrame by a specific cell stage
so that there is a "common" reference point (like common start time of lineage)
}
