% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEmb.R
\name{drawEmbLine}
\alias{drawEmbLine}
\title{drawEmbLine draw the nucleus positions of the embryo in 3D scatter plots, colored by lineage}
\usage{
drawEmbLine(
  embryoCD,
  time,
  lineages = NULL,
  colors = NULL,
  opacity_s = NULL,
  xSize = 1,
  ySize = 1,
  zSize = 1,
  otherOpacity = 0.2,
  cellSize = 7.5,
  center = list(x = 0, y = 0, z = 0),
  viewPoint = list(x = 0, y = 0, z = 1.8),
  showAxis = F
)
}
\arguments{
\item{embryoCD}{the embryo dataframe with cell, time column, mandatory}

\item{time}{which time to plot the embryo, mandatory}

\item{lineages}{a list of lineages to highlight (plot with the given color and opacity in "colors" and "opacity_s" parameter)}

\item{colors}{a list of colors to plot each lineage given in "lineages"}

\item{opacity_s}{a list of opacity for each lineage given in "lineages"}

\item{xSize}{size of each x unit in embryoCD dataframe}

\item{ySize}{size of each y unit in embryoCD dataframe}

\item{zSize}{size of each z unit in embryoCD dataframe}

\item{otherOpacity}{opacity for cells not specified}

\item{cellSize}{single numeric, size of the data points}

\item{center}{which point in the 3d space does the camera aim at}

\item{viewPoint}{where to put the imaginative camera initially}

\item{showAxis}{wherther to diaply axis and axis labels}
}
\value{
a list made of: the plotly figure, a dataframe of the highlighted cells, and a dataframe of cells not highlighted
}
\description{
drawEmbLine draw the nucleus positions of the embryo in 3D scatter plots, colored by lineage
}
\examples{
drawEmbVal(Embryo, lineages = c("MS", "E", "C"), colors = c(rgb(1,0,0), rgb(1,0,1), rgb(0,0,1)), time = 139)
}
