% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEmb.R
\name{drawEmbLine}
\alias{drawEmbLine}
\title{drawEmbLine draw the nucleus positions of the embryo in 3D scatter plots, colored by lineage}
\usage{
drawEmbLine(
  embryoCD,
  time,
  lineages = NULL,
  colors = NULL,
  opacity_s = NULL,
  xSize = 1,
  ySize = 1,
  zSize = 1,
  aligned = F,
  otherOpacity = 0.2,
  cellSize = 7.5,
  center = list(x = 0, y = 0, z = 0),
  viewPoint = list(x = 0, y = 0, z = 1.8)
)
}
\arguments{
\item{embryoCD}{the embryo dataframe with cell, time column, mandatory}

\item{time}{numeric, mandatory, which time to plot the embryo, automatically choose the closest time point if no exact match}

\item{lineages}{a list of lineages to highlight (plot with the given color and opacity in "colors" and "opacity_s" parameter)}

\item{colors}{a list of colors to plot each lineage given in "lineages"}

\item{opacity_s}{a list of opacity for each lineage given in "lineages"}

\item{xSize}{size of each x unit in embryoCD dataframe, default 1}

\item{ySize}{size of each y unit in embryoCD dataframe, default 1}

\item{zSize}{size of each z unit in embryoCD dataframe, default 1}

\item{otherOpacity}{opacity for cells not specified}

\item{cellSize}{single numeric, size of the data points}

\item{center}{A named list with elements:
\describe{
\item{x}{Numeric. X coordinate.}
\item{y}{Numeric. Y coordinate.}
\item{z}{Numeric. Z coordinate.}
}
which point in the 3d space does the camera aim at, default list(x=0,y=0,z=0)}

\item{viewPoint}{A named list with elements:
\describe{
\item{x}{Numeric. X coordinate.}
\item{y}{Numeric. Y coordinate.}
\item{z}{Numeric. Z coordinate.}
}
where to put the imaginative camera initially, default list(x=0,y=0,z=1.8)}
}
\value{
a list made of: the plotly figure, a dataframe of the highlighted cells, and a dataframe of cells not highlighted
}
\description{
drawEmbLine draw the nucleus positions of the embryo in 3D scatter plots, colored by lineage
}
\examples{
drawEmbVal(Embryo, lineages = c("MS", "E", "C"), colors = c(rgb(1,0,0), rgb(1,0,1), rgb(0,0,1)), time = 139)
}
