% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEmb.R
\name{drawEmbVal}
\alias{drawEmbVal}
\title{drawEmbVal}
\usage{
drawEmbVal(
  embryoCD,
  time,
  valCol = "blot",
  title = NULL,
  lineages = NULL,
  shapes = NULL,
  cellSize = 10,
  ReporterForAll = F,
  colorScheme = NULL,
  alpha_selected = NULL,
  alpha_other = NULL,
  maxBlot = NULL,
  minBlot = NULL,
  xSize = 1,
  ySize = 1,
  zSize = 1,
  aligned = F,
  center = list(x = 0, y = 0, z = 0),
  viewPoint = list(x = 0, y = 0, z = 1.8)
)
}
\arguments{
\item{embryoCD}{the embryo dataframe with cell, time, and blot(expression) column, mandatory}

\item{time}{which time to plot the embryo, mandatory}

\item{valCol}{column that will color the nucleus by}

\item{title}{figure title text}

\item{lineages}{a list of lineages to highlight (plot with the given symbol in "shapes" parameter)}

\item{shapes}{a list of shapes to plot each lineage given in "lineages"}

\item{ReporterForAll}{color the nucleus points by expression or not (default TRUE)}

\item{colorScheme}{A color scale definition. This can be:
\itemize{
\item \strong{A list of vectors:} Each sub-vector maps a normalized value
(0 to 1) to an R color.
\itemize{
\item Must include mappings for 0 and 1. Example: \code{list(c(0, "blue"), c(1, "red"))}
\item Can add intermediate values: \code{list(c(0, "blue"), c(0.5, "white"), c(1, "red"))}
}
\item \strong{A palette name string:} One of the
\href{https://plotly.com/r/reference/scatter3d/}{Plotly scatter3d} compatible strings:
\emph{Blackbody, Bluered, Blues, Cividis, Earth, Electric, Greens, Greys,
Hot, Jet, Picnic, Portland, Rainbow, RdBu, Reds, Viridis, YlGnBu, YlOrRd.}
}}

\item{alpha_selected}{alpha value (opacity) for the selected lineages}

\item{alpha_other}{alpha value (opacity) for other lineages}

\item{maxBlot}{upper limit for expression coloring}

\item{minBlot}{lower limit for expresion coloring}

\item{xSize}{size of each x unit in embryoCD dataframe, default 1}

\item{ySize}{size of each y unit in embryoCD dataframe, default 1}

\item{zSize}{size of each z unit in embryoCD dataframe, default 1}

\item{center}{A named list with elements:
\describe{
\item{x}{Numeric. X coordinate.}
\item{y}{Numeric. Y coordinate.}
\item{z}{Numeric. Z coordinate.}
}
which point in the 3d space does the camera aim at, default list(x=0,y=0,z=0)}

\item{viewPoint}{A named list with elements:
\describe{
\item{x}{Numeric. X coordinate.}
\item{y}{Numeric. Y coordinate.}
\item{z}{Numeric. Z coordinate.}
}
where to put the imaginative camera initially, default list(x=0,y=0,z=1.8)}
}
\value{
a list made of: the plotly figure, a dataframe of the highlighted cells, and a dataframe of cells not highlighted
}
\description{
draw the nucleus positions of the embryo in 3D scatter plots, colored by expression
}
