% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEmb.R
\name{drawEmbVal}
\alias{drawEmbVal}
\title{drawEmb
This file contains several functions relevant to drawing a 3D (interactive) embryo,
able to represent expression levels and highlight lineages
if deliver multiple "lineages" must also deliver a list "lineageColors" with elements corresponding to "lineages"
drawEmbVal draw the nucleus positions of the embryo in 3D scatter plots, colored by expression}
\usage{
drawEmbVal(
  embryoCD,
  time,
  valCol = "blot",
  lineages = NULL,
  shapes = NULL,
  cellSize = 10,
  ReporterForAll = F,
  colorScheme = NULL,
  alpha_selected = NULL,
  alpha_other = NULL,
  maxBlot = NULL,
  minBlot = NULL,
  xSize = 1,
  ySize = 1,
  zSize = 1,
  aligned = F,
  center = list(x = 0, y = 0, z = 0),
  viewPoint = list(x = 0, y = 0, z = 1.8)
)
}
\arguments{
\item{embryoCD}{the embryo dataframe with cell, time, and blot(expression) column, mandatory}

\item{time}{which time to plot the embryo, mandatory}

\item{valCol}{column that will color the nucleus by}

\item{lineages}{a list of lineages to highlight (plot with the given symbol in "shapes" parameter)}

\item{shapes}{a list of shapes to plot each lineage given in "lineages"}

\item{ReporterForAll}{color the nucleus points by expression or not (default TRUE)}

\item{colorScheme}{a list of at least 2 vectors made of a number between 0 and 1 to specify expression fraction, and a color value to specify color fro that fraction}

\item{alpha_selected}{alpha value (opacity) for the selected lineages}

\item{alpha_other}{alpha value (opacity) for other lineages}

\item{maxBlot}{upper limit for expression coloring}

\item{minBlot}{lower limit for expresion coloring}

\item{xSize}{size of each x unit in embryoCD dataframe, default 1}

\item{ySize}{size of each y unit in embryoCD dataframe, default 1}

\item{zSize}{size of each z unit in embryoCD dataframe, default 1}

\item{center}{A named list with elements:
\describe{
\item{x}{Numeric. X coordinate.}
\item{y}{Numeric. Y coordinate.}
\item{z}{Numeric. Z coordinate.}
}
which point in the 3d space does the camera aim at, default list(x=0,y=0,z=0)}

\item{viewPoint}{A named list with elements:
\describe{
\item{x}{Numeric. X coordinate.}
\item{y}{Numeric. Y coordinate.}
\item{z}{Numeric. Z coordinate.}
}
where to put the imaginative camera initially, default list(x=0,y=0,z=1.8)}
}
\value{
a list made of: the plotly figure, a dataframe of the highlighted cells, and a dataframe of cells not highlighted
}
\description{
drawEmb
This file contains several functions relevant to drawing a 3D (interactive) embryo,
able to represent expression levels and highlight lineages
if deliver multiple "lineages" must also deliver a list "lineageColors" with elements corresponding to "lineages"
drawEmbVal draw the nucleus positions of the embryo in 3D scatter plots, colored by expression
}
