% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Processing.R
\name{plotBlotLine}
\alias{plotBlotLine}
\title{plotBlotLine}
\usage{
plotBlotLine(
  CDDataL,
  title,
  aligningCell,
  align_t = "mean",
  cells = NULL,
  lineages = NULL,
  excludeCD = NULL,
  colorGroup = "name",
  color_palette = NULL,
  xrange = list(NULL, NULL),
  yrange = list(NULL, NULL)
)
}
\arguments{
\item{CDDataL}{the list of CD Data generated by function \code{CD_In}}

\item{cells}{(list of) exact cell names to extract}

\item{lineages}{(list of) mother cell names of lineages to extract}

\item{excludeCD}{(list of) CD names to exclude}

\item{colorGroup}{the embryo attribute by which we assign color to its corresponding figure curve, use embryo name by default}

\item{color_palette}{a named vector that map each possible colorGroup value to a color, use NULL to autogenerate a color palette by colorGroup}

\item{xrange}{a list in format list(min,max) to define range of x axis}

\item{yrange}{a list in format list(min,max) to define range of y axis}

\item{alignCell}{the cell used to find common time point to align across embryos}
}
\value{
a single plotly line plot with each line the (average) blot vs time for each embryo
}
\description{
plots the averaged blot value across each timepoint for selected cells/lineages from all CD data available
}
