% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageProcessing.R
\name{RePosition}
\alias{RePosition}
\title{RePosition}
\usage{
RePosition(
  CDFrame,
  time,
  indicatorP = "C",
  indicatorD = "Cxa",
  indicatorV = "MSxxp",
  indicatorL = NULL,
  indicatorR = NULL,
  xSize = 0.08651,
  ySize = 0.08651,
  zSize = 0.5
)
}
\arguments{
\item{CDFrame}{a CD-like dataFrame}

\item{time}{the time to align}

\item{indicatorP}{lineage on the relative posterior of embryo center}

\item{indicatorD}{lineage on the relative dorsal of embryo center, use NULL not NaN if not considering}

\item{indicatorV}{lineage on the relative ventral of embryo center, use NULL not NaN if not considering}

\item{indicatorL}{lineage on the relative left of embryo center, use NULL not NaN if not considering}

\item{indicatorR}{lineage on the relative right of embryo center, use NULL not NaN if not considering}

\item{xSize}{size of each unit on x-axis in original CDFrame}

\item{ySize}{size of each unit on y-axis in original CDFrame}

\item{zSize}{size of each unit on z-axis in original CDFrame}
}
\value{
the slice of the provided CDFrame at time==time, rotated
}
\description{
orientate one embryo at a given time point so that
the embryo is centered at (0,0,0)
coordinate the AP axis is aligned to the x-axis,
DV to z-axis and LR to y-axis.
Note that other time points are not adjusted - use totalRePosition to align all time points
}
